requirements.txt
nats-py
pyyaml
bson
